refactor: reuse as much as possible
