+++ src/queries/chats.web.ts
@@ -1,6 +1,4 @@
-import {useSuspenseQuery} from '@tanstack/react-query';
 import type {Chat} from '../lib/types';
-import {chatsKeys} from '../lib/keys';
 
 interface ChatWithLastMessage extends Chat {
   last_message: string;
@@ -16,13 +14,4 @@ export const getRecentChats = async () => {
       last_message_author_email: 'placeholder',
     },
   ] as ChatWithLastMessage[]);
-};
-
-export const useRecentChats = () => {
-  const recentChats = useSuspenseQuery({
-    queryKey: chatsKeys.all,
-    queryFn: getRecentChats,
-  });
-
-  return recentChats;
 };
