+++ src/queries/kv.ts
@@ -1,6 +1,4 @@
-import {useSuspenseQuery} from '@tanstack/react-query';
 import {db} from '../lib/db';
-import type {Chat} from '../lib/types';
 
 export const getByKey = async <T>(key: string) => {
   const response = await db.executeAsync(
@@ -11,15 +9,3 @@ export const getByKey = async <T>(key: string) => {
   // TODO: handle json parsing with sqlite
   return JSON.parse(response.rows?._array[0].value) as T;
 };
-
-export const useKey = <T>(key: string) => {
-  const query = useSuspenseQuery({
-    queryKey: [key],
-    queryFn: () => getByKey<T>(key),
-  });
-
-  return query;
-};
-
-export const useBetas = () => useKey<string[]>('betas');
-export const useChat = (chatId: number) => useKey<Chat>(`chat_${chatId}`);
